Refactor ProductImage so that ProductThumbnail logic is in its own component
