+++ client/src/ProductView/ProductImage.jsx
@@ -2,13 +2,10 @@ import React, { useState } from 'react';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import ReactImageMagnify from 'react-image-magnify';
-
-// TODO: move thumbnails logic to different component
-// TODO: make sure active class is still added to thumbnail when arrows are used to toggle between images
+import ProductThumbnails from './ProductThumbnails.jsx';
 
 const ProductImage = (props) => {
   const [currentImage, setCurrentImage] = useState(0);
-  let thumbnailsArrow = false;
 
   const removeActiveClass = (e) => {
     let thumbnails = document.querySelectorAll('.thumbnail');
@@ -25,11 +22,6 @@ const ProductImage = (props) => {
     removeActiveClass(e);
   };
 
-  const handleThumbnailArrowClick = () => {
-    let thumbnailContainer = document.querySelector('.thumbnails');
-    thumbnailContainer.scrollTo({ top: 600, behavior: 'smooth' })
-  };
-
   const handleArrowClick = (direction) => {
     if (direction === 'right') {
       setCurrentImage(currentImage + 1)
@@ -41,28 +33,10 @@ const ProductImage = (props) => {
   return (
     <Col sm={8} className="imageContainer">
       <Row>
-        <Col sm={2} className="thumbnails">
-          { props.productStyle
-            ? props.productStyle.photos.map((photo, index) => {
-                if (index > 6) {
-                  thumbnailsArrow = true;
-                }
-                return <div
-                  className="thumbnail"
-                  style={{ backgroundImage: `url(${photo.thumbnail_url})` }}
-                  key={index}
-                  onClick={(e) => handleThumbnailClick(e, index)}
-                  ></div>
-              })
-            : <p>Loading...</p>
-          }
-        </Col>
-        { thumbnailsArrow
-          ? <p className="arrow" onClick={handleThumbnailArrowClick}>
-              <i className="fa fa-arrow-circle-o-down" aria-hidden="true"></i>
-            </p>
-          : ''
-        }
+        <ProductThumbnails
+          productStyle={props.productStyle}
+          handleThumbnailClick={handleThumbnailClick} />
+
         <Col sm={10} className="mainImage fluid">
           <i className="fa fa-arrow-left"
             aria-hidden="true"
@@ -79,7 +53,7 @@ const ProductImage = (props) => {
                 largeImage: {
                   src: props.productStyle.photos[currentImage].url,
                   width: 1200,
-                  height: 1800,
+                  height: 1000,
                 },
                 isHintEnabled: true
               }} />
