+++ client/src/ProductView/StyleSelector.jsx
@@ -5,7 +5,7 @@ const StyleSelector = (props) => {
   return (
     <div>
       <p className="currentStyle">Style > <b>{props.styleDetails.name}</b></p>
-      {props.allStyles && props.allStyles.length
+      { props.allStyles && props.allStyles.length
           ? props.allStyles.map(style => {
               return <p>{style.name}</p>;
            })
